requirements.txt
openai-whisper>=20231117
deep-translator>=1.11.4
edge-tts>=6.1.9
moviepy==1.0.3
torch>=2.0.0
torchaudio>=2.0.0
streamlit>=1.28.0

